requirements.txt
flask==3.0.0
flask-talisman==1.0.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
face-recognition==1.3.0
gunicorn==21.2.0
Pillow==10.0.0
